Float values works just fine - no need to use double

+++ source/profdump/func.d
@@ -30,9 +30,9 @@ struct Function {
 		this.CalledBy[h] = func;
 	}
 
-	const void toString(scope void delegate(const(char)[]) s, ulong tps = 0, double threshold = 0) {
+	const void toString(scope void delegate(const(char)[]) s, ulong tps = 0, float threshold = 0) {
 		import std.format : format;
-		if(threshold != 0 && (cast(double) this.Time / cast(double) tps) < threshold)
+		if(threshold != 0 && (cast(float) this.Time / cast(float) tps) < threshold)
 			return;
 		s("Function '%s':\n".format(this.Name));
 		s("\tMangled name: '%s'\n".format(this.Mangled));
@@ -48,9 +48,9 @@ struct Function {
 		}
 		if(tps) {
 			s("\tFinished in: %f seconds (just this function)\n"
-				.format(cast(double) this.FunctionTime / cast(double) tps));
+				.format(cast(float) this.FunctionTime / cast(float) tps));
 			s("\tFinished in: %f seconds (this function and all descendants)\n"
-				.format(cast(double) this.Time / cast(double) tps));
+				.format(cast(float) this.Time / cast(float) tps));
 		} else {
 			s("\tFinished in: %d ticks (just this function)\n".format(this.FunctionTime));
 			s("\tFinished in: %d ticks (this function and all descendants)\n"
@@ -58,8 +58,8 @@ struct Function {
 		}
 	}
 
-	JSONValue toJSON(ulong tps = 0, double threshold = 0) {
-		if(threshold != 0 && (cast(double) this.Time / cast(double) tps) < threshold)
+	JSONValue toJSON(ulong tps = 0, float threshold = 0) {
+		if(threshold != 0 && (cast(float) this.Time / cast(float) tps) < threshold)
 			return JSONValue(null);
 		JSONValue ret = JSONValue([
 			"name": this.Name,
@@ -89,9 +89,9 @@ struct Function {
 		}
 		if(tps) {
 			ret["functionTimeSec"] = JSONValue(
-				cast(double) this.FunctionTime / cast(double) tps);
+				cast(float) this.FunctionTime / cast(float) tps);
 			ret["timeSec"] = JSONValue(
-				cast(double) this.Time / cast(double) tps);
+				cast(float) this.Time / cast(float) tps);
 		}
 		ret["functionTime"] = JSONValue(this.FunctionTime);
 		ret["time"] = JSONValue(this.Time);
